Review comments incorporated
