+++ utils/utils.js
@@ -61,3 +61,18 @@ export function getIcon(tag) {
 
   return img;
 }
+
+export function addRepeater(element, title) {
+  element.lastChild.setAttribute('repeatIdx', 0);
+
+  const tag = createTag('div');
+  tag.classList.add('repeater-element');
+
+  const heading = createTag('h3', { class: 'repeater-element-title' }, title);
+  tag.append(heading);
+
+  const plusIcon = getIcon('add-circle');
+  tag.append(plusIcon);
+
+  element.append(tag);
+}
